Extract out RequestorPageContent
