+++ studio/src/pages/RequestorPage/RequestorPage.tsx
@@ -16,6 +16,7 @@ import { useParams } from "react-router-dom";
 import { NavigationPanel } from "./NavigationPanel";
 import { RequestPanel } from "./RequestPanel";
 import { RequestorInput } from "./RequestorInput";
+import { RequestorPageContent } from "./RequestorPageContent";
 import { ResponsePanel } from "./ResponsePanel";
 import { RoutesCombobox } from "./RoutesCombobox";
 import { AiTestGenerationPanel, useAi } from "./ai";
@@ -35,13 +36,7 @@ function getMainSectionWidth() {
   return window.innerWidth - 400;
 }
 
-function getMainSectionHeight() {
-  return window.innerHeight - 150;
-}
-
 export const RequestorPage = () => {
-  const { toast } = useToast();
-
   const { id } = useParams();
   // NOTE - This sets the `routes` and `serviceBaseUrl` in the reducer
   useRoutes();
@@ -63,87 +58,11 @@ export const RequestorPage = () => {
   } = useRequestorHistory();
 
   const hasHistory = history.length > 0;
-  console.log("hasHistory", history);
   useEffect(() => {
     if (id && hasHistory) {
-      console.log("id", id);
       loadHistoricalRequest(id);
     }
   }, [id, loadHistoricalRequest, hasHistory]);
-
-  const mostRecentRequestornatorForRoute =
-    useMostRecentRequestornator(sessionHistory);
-
-  const { mutate: makeRequest, isPending: isRequestorRequesting } =
-    useMakeProxiedRequest();
-
-  // WIP - Allows us to connect to a websocket and send messages through it
-  const {
-    connect: connectWebsocket,
-    disconnect: disconnectWebsocket,
-    sendMessage: sendWebsocketMessage,
-    state: websocketState,
-  } = useMakeWebsocketRequest();
-
-  // Send a request when we submit the form
-  const onSubmit = useRequestorSubmitHandler({
-    makeRequest,
-    connectWebsocket,
-    recordRequestInSessionHistory,
-  });
-
-  const formRef = useRef<HTMLFormElement>(null);
-
-  // FIXME / INVESTIGATE - Should this behavior change for websockets?
-  useHotkeys(
-    "mod+enter",
-    () => {
-      if (formRef.current) {
-        formRef.current.requestSubmit();
-      }
-    },
-    {
-      enableOnFormTags: ["input"],
-    },
-  );
-
-  const {
-    enabled: aiEnabled,
-    isLoadingParameters,
-    fillInRequest,
-    testingPersona,
-    setTestingPersona,
-    showAiGeneratedInputsBanner,
-    setShowAiGeneratedInputsBanner,
-    setIgnoreAiInputsBanner,
-  } = useAi(history);
-
-  useHotkeys(
-    "mod+g",
-    (e) => {
-      if (aiEnabled) {
-        // Prevent the "find in document" from opening in browser
-        e.preventDefault();
-        if (!isLoadingParameters) {
-          toast({
-            duration: 3000,
-            description: "Generating request parameters with AI",
-          });
-          fillInRequest();
-        }
-      }
-    },
-    {
-      enableOnFormTags: ["input"],
-    },
-  );
-
-  const [isAiTestGenerationPanelOpen, setIsAiTestGenerationPanelOpen] =
-    useState(false);
-  const toggleAiTestGenerationPanel = useCallback(
-    () => setIsAiTestGenerationPanelOpen((current) => !current),
-    [],
-  );
 
   const width = getMainSectionWidth();
   const isLgScreen = useIsLgScreen();
@@ -155,45 +74,9 @@ export const RequestorPage = () => {
     minimalGroupSize: 944,
   });
 
-  const { minSize: requestPanelMinSize, maxSize: requestPanelMaxSize } =
-    usePanelConstraints({
-      // Change the groupId to `""` on small screens because we're not rendering
-      // the resizable panel group
-      groupId: "requestor-page-main-panel",
-      initialGroupSize: getMainSectionHeight(),
-      minPixelSize: 200,
-      dimension: "height",
-    });
-
-  const requestContent = (
-    <RequestPanel
-      aiEnabled={aiEnabled}
-      isLoadingParameters={isLoadingParameters}
-      fillInRequest={fillInRequest}
-      testingPersona={testingPersona}
-      setTestingPersona={setTestingPersona}
-      showAiGeneratedInputsBanner={showAiGeneratedInputsBanner}
-      setShowAiGeneratedInputsBanner={setShowAiGeneratedInputsBanner}
-      setIgnoreAiInputsBanner={setIgnoreAiInputsBanner}
-      websocketState={websocketState}
-      sendWebsocketMessage={sendWebsocketMessage}
-    />
-  );
-
-  const responseContent = (
-    <ResponsePanel
-      tracedResponse={mostRecentRequestornatorForRoute}
-      isLoading={isRequestorRequesting}
-      websocketState={websocketState}
-      openAiTestGenerationPanel={toggleAiTestGenerationPanel}
-      isAiTestGenerationPanelOpen={isAiTestGenerationPanelOpen}
-    />
-  );
-
   return (
     <div
       className={cn(
-        // It's critical the parent has a fixed height for our grid layout to work
         "h-[calc(100vh-40px)]",
         "flex",
         "flex-col",
@@ -203,16 +86,7 @@ export const RequestorPage = () => {
         "lg:gap-4",
       )}
     >
-      <div
-        className={cn(
-          "relative",
-          "lg:overflow-y-auto",
-          "lg:overflow-x-hidden",
-          "lg:hidden",
-        )}
-      >
-        <RoutesCombobox />
-      </div>
+      {/* ... existing code ... */}
       <ResizablePanelGroup
         direction="horizontal"
         id="requestor-page-main"
@@ -237,101 +111,23 @@ export const RequestorPage = () => {
           </>
         )}
         <ResizablePanel id="main" order={1}>
-          <div
-            className={cn(
-              "flex",
-              "flex-col",
-              "gap-2",
-              // HACK - This is a workaround to prevent the grid from overflowing on smaller screens
-              "h-[calc(100%-0.6rem)]",
-              "lg:h-full",
-              "relative",
-              // "overflow-scroll",
-              "overflow-hidden",
-            )}
-          >
-            <RequestorInput
-              onSubmit={onSubmit}
-              disconnectWebsocket={disconnectWebsocket}
-              isRequestorRequesting={isRequestorRequesting}
-              formRef={formRef}
-              websocketState={websocketState}
-            />
-            <ResizablePanelGroup
-              direction="vertical"
-              id="requestor-page-main-panel"
-              autoSaveId="requestor-page-main-panel"
-            >
-              <ResizablePanel
-                defaultSize={isAiTestGenerationPanelOpen ? 50 : 100}
-              >
-                <ResizablePanelGroup
-                  direction={isLgScreen ? "horizontal" : "vertical"}
-                  id="requestor-page-request-panel-group"
-                  autoSaveId="requestor-page-request-panel-group"
-                  className={cn(
-                    "rounded-md",
-                    // HACK - This defensively prevents overflow from getting too excessive,
-                    //        In the case where the inner content expands beyond the parent
-                    "max-w-screen",
-                    "max-h-full",
-                    // "gap-1",
-                  )}
-                >
-                  <ResizablePanel
-                    order={1}
-                    className={cn(BACKGROUND_LAYER, "relative")}
-                    id="request-panel"
-                    minSize={requestPanelMinSize}
-                    maxSize={requestPanelMaxSize}
-                  >
-                    {requestContent}
-                  </ResizablePanel>
-                  <ResizableHandle
-                    hitAreaMargins={{ coarse: 20, fine: 10 }}
-                    className="bg-transparent"
-                  />
-                  <ResizablePanel
-                    id="response-panel"
-                    order={4}
-                    minSize={10}
-                    className={cn(BACKGROUND_LAYER)}
-                  >
-                    {responseContent}
-                  </ResizablePanel>
-                </ResizablePanelGroup>
-              </ResizablePanel>
-              <ResizablePanel>
-                {isAiTestGenerationPanelOpen && (
-                  <>
-                    <ResizableHandle
-                      hitAreaMargins={{ coarse: 20, fine: 10 }}
-                      className="bg-transparent"
-                    />
-                    <ResizablePanel
-                      order={3}
-                      id="ai-panel"
-                      className={cn(
-                        BACKGROUND_LAYER,
-                        "rounded-md",
-                        "border",
-                        "h-full",
-                        "mt-2",
-                      )}
-                    >
-                      <AiTestGenerationPanel
-                        // TODO - Only use history for recent matching route
-                        history={history}
-                        toggleAiTestGenerationPanel={
-                          toggleAiTestGenerationPanel
-                        }
-                      />
-                    </ResizablePanel>
-                  </>
-                )}
-              </ResizablePanel>
-            </ResizablePanelGroup>
-          </div>
+          <RequestorPageContent
+            // onSubmit={onSubmit}
+            // disconnectWebsocket={disconnectWebsocket}
+            // isRequestorRequesting={isRequestorRequesting}
+            // formRef={formRef}
+            // websocketState={websocketState}
+            // isLgScreen={isLgScreen}
+            // requestPanelMinSize={requestPanelMinSize}
+            // requestPanelMaxSize={requestPanelMaxSize}
+            // requestContent={requestContent}
+            // responseContent={responseContent}
+            // isAiTestGenerationPanelOpen={isAiTestGenerationPanelOpen}
+            // toggleAiTestGenerationPanel={toggleAiTestGenerationPanel}
+            history={history}
+            sessionHistory={sessionHistory}
+            recordRequestInSessionHistory={recordRequestInSessionHistory}
+          />
         </ResizablePanel>
       </ResizablePanelGroup>
     </div>
@@ -339,55 +135,6 @@ export const RequestorPage = () => {
 };
 
 export default RequestorPage;
-
-/**
- * When you select a route from the route side panel,
- * this will look for the most recent request made against that route.
- */
-function useMostRecentRequestornator(all: Requestornator[]) {
-  const { path: routePath } = useActiveRoute();
-  const { path, method, activeHistoryResponseTraceId } = useRequestorStore(
-    "path",
-    "method",
-    "activeHistoryResponseTraceId",
-  );
-  return useMemo<Requestornator | undefined>(() => {
-    if (activeHistoryResponseTraceId) {
-      return all.find(
-        (r: Requestornator) =>
-          r?.app_responses?.traceId === activeHistoryResponseTraceId,
-      );
-    }
-
-    const matchingResponses = all?.filter(
-      (r: Requestornator) =>
-        r?.app_requests?.requestRoute === routePath &&
-        r?.app_requests?.requestMethod === method,
-    );
-
-    // Descending sort by updatedAt
-    matchingResponses?.sort(sortRequestornatorsDescending);
-
-    const latestMatch = matchingResponses?.[0];
-
-    if (latestMatch) {
-      return latestMatch;
-    }
-
-    // HACK - We can try to match against the exact request URL
-    //        This is a fallback to support the case where the route doesn't exist,
-    //        perhaps because we made a request to a service we are not explicitly monitoring
-    const matchingResponsesFallback = all?.filter(
-      (r: Requestornator) =>
-        r?.app_requests?.requestUrl === path &&
-        r?.app_requests?.requestMethod === method,
-    );
-
-    matchingResponsesFallback?.sort(sortRequestornatorsDescending);
-
-    return matchingResponsesFallback?.[0];
-  }, [all, routePath, method, path, activeHistoryResponseTraceId]);
-}
 
 export const Title = (props: { children: React.ReactNode }) => (
   <div
